WalletService provided for the store
                I've updated the store getters in order to obtain the
                walletService instance to be used. Also, I've updated
                the types and utils imports so that we can prevent
                dependency cycles.
