+++ src/components/create/BtcInputAmount.vue
@@ -69,7 +69,7 @@ import {
 } from 'vue-property-decorator';
 import { Action, State } from 'vuex-class';
 import SatoshiBig from '@/types/SatoshiBig';
-import { PegInTxState } from '@/store/peginTx/types';
+import { PegInTxState } from '@/types/pegInTx';
 import * as constants from '@/store/constants';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
