+++ src/components/create/BtcTxSummarySide.vue
@@ -129,8 +129,8 @@ import {
 } from 'vue-property-decorator';
 import { Getter, State } from 'vuex-class';
 import Big from 'big.js';
-import { PegInTxState } from '@/store/peginTx/types';
-import { getChunkedValue } from '@/utils';
+import { PegInTxState } from '@/types/pegInTx';
+import { getChunkedValue } from '@/utils/common';
 import SatoshiBig from '@/types/SatoshiBig';
 import * as constants from '@/store/constants';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
