+++ src/services/TrezorService.ts
@@ -1,12 +1,12 @@
 import TrezorConnect, { GetAddress } from 'trezor-connect';
 import * as bitcoin from 'bitcoinjs-lib';
 import { Network } from 'bitcoinjs-lib';
-import { WalletAddress } from '@/store/peginTx/types';
+import { WalletAddress } from '@/types/pegInTx';
 import * as constants from '@/store/constants';
 import {
   TrezorSignedTx, TrezorTx, Tx,
 } from '@/types';
-import { WalletService } from '@/services/WalletService';
+import { WalletService } from '@/services';
 import { EnvironmentAccessorService } from './enviroment-accessor.service';
 
 export default class TrezorService extends WalletService {
