+++ src/services/index.ts
@@ -1,0 +1,8 @@
+// eslint-disable-next-line import/no-cycle
+export { default as WalletService } from './WalletService';
+// eslint-disable-next-line import/no-cycle
+export { default as ApiService } from './ApiService';
+// eslint-disable-next-line import/no-cycle
+export { default as TrezorService } from './TrezorService';
+// eslint-disable-next-line import/no-cycle
+export { default as LedgerService } from './LedgerService';
