+++ src/store/session/actions.ts
@@ -5,7 +5,7 @@ import RLogin from '@rsksmart/rlogin';
 import WalletConnectProvider from '@walletconnect/web3-provider';
 import * as constants from '@/store/constants';
 import { TransactionType, SessionState } from './types';
-import { RootState } from '../types';
+import { RootState } from '../../types/store';
 
 export const actions: ActionTree<SessionState, RootState> = {
   [constants.SESSION_CONNECT_WEB3]: ({ commit, state }): Promise<void> => {
