+++ src/store/session/getters.ts
@@ -1,7 +1,7 @@
 import { GetterTree } from 'vuex';
 import * as constants from '@/store/constants';
 import { SessionState } from './types';
-import { RootState } from '../types';
+import { RootState } from '../../types/store';
 
 export const getters: GetterTree<SessionState, RootState> = {
   [constants.SESSION_IN_TX_FLOW]: (state): boolean => state.txType !== undefined,
