+++ src/types/index.ts
@@ -2,3 +2,5 @@ export * from './Common';
 export * from './Trezor';
 export * from './Ledger';
 export * from './Wallets';
+export * from './pegInTx';
+export * from './store';
