+++ src/types/store.ts
