+++ src/views/Exchange.vue
@@ -15,7 +15,7 @@ import SendBitcoinLedger from '@/components/ledger/SendBitcoinLedger.vue';
 import SuccessSend from '@/components/exchange/SuccessSend.vue';
 import TrackingId from '@/components/exchange/TrackingId.vue';
 import * as constants from '@/store/constants';
-import { PegInTxState } from '@/store/peginTx/types';
+import { PegInTxState } from '@/types/pegInTx';
 
 @Component({
   components: {
