+++ src/views/Status.vue
@@ -173,8 +173,8 @@ import { State } from 'vuex-class';
 import TxSummary from '@/components/exchange/TxSummary.vue';
 import { PegStatus } from '@/store/constants';
 import ApiService from '@/services/ApiService';
-import { PeginStatus } from '@/store/types';
-import { PegInTxState } from '@/store/peginTx/types';
+import { PeginStatus } from '@/types/store';
+import { PegInTxState } from '@/types/pegInTx';
 import { TxData } from '@/types';
 import SatoshiBig from '@/types/SatoshiBig';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
